[MOD/#255] 바텀시트 내용 수정
