+++ presentation/diaryfeedback/src/main/java/com/hilingual/presentation/diaryfeedback/component/FeedbackMenuBottomSheet.kt
@@ -1,0 +1,76 @@
+/*
+ * Copyright 2025 The Hilingual Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.hilingual.presentation.diaryfeedback.component
+
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import com.hilingual.core.designsystem.component.bottomsheet.HilingualMenuBottomSheet
+import com.hilingual.core.designsystem.component.bottomsheet.HilingualMenuBottomSheetItem
+import com.hilingual.core.designsystem.theme.HilingualTheme
+import com.hilingual.core.designsystem.R as DesignSystemR
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+internal fun FeedbackMenuBottomSheet(
+    isVisible: Boolean,
+    onDismiss: () -> Unit,
+    onDeleteClick: () -> Unit,
+    onReportClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    HilingualMenuBottomSheet(
+        isVisible = isVisible,
+        onDismiss = onDismiss,
+        modifier = modifier
+    ) {
+        HilingualMenuBottomSheetItem(
+            text = "삭제하기",
+            iconResId = DesignSystemR.drawable.ic_delete_24,
+            onClick = onDeleteClick,
+            textColor = HilingualTheme.colors.alertRed
+        )
+
+        HilingualMenuBottomSheetItem(
+            text = "AI 피드백 신고하기",
+            iconResId = DesignSystemR.drawable.ic_report_24,
+            onClick = onReportClick
+        )
+    }
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Preview(showBackground = true)
+@Composable
+private fun FeedbackReportPreview() {
+    var isVisible by remember { mutableStateOf(true) }
+
+    HilingualTheme {
+        HilingualTheme {
+            FeedbackMenuBottomSheet(
+                isVisible = isVisible,
+                onDismiss = { isVisible = false },
+                onDeleteClick = {},
+                onReportClick = {}
+            )
+        }
+    }
+}
