--- presentation/diaryfeedback/src/main/java/com/hilingual/presentation/diaryfeedback/component/FeedbackReportBottomSheet.kt
@@ -1,115 +1,0 @@
-/*
- * Copyright 2025 The Hilingual Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hilingual.presentation.diaryfeedback.component
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.hilingual.core.common.extension.noRippleClickable
-import com.hilingual.core.designsystem.component.bottomsheet.HilingualBasicBottomSheet
-import com.hilingual.core.designsystem.theme.HilingualTheme
-import com.hilingual.core.designsystem.R as DesignSystemR
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-internal fun FeedbackReportBottomSheet(
-    isVisible: Boolean,
-    onDismiss: () -> Unit,
-    onReportClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    HilingualBasicBottomSheet(
-        isVisible = isVisible,
-        onDismiss = onDismiss,
-        modifier = modifier
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(vertical = 24.dp, horizontal = 16.dp)
-        ) {
-            Text(
-                text = "AI 피드백",
-                style = HilingualTheme.typography.headB16,
-                color = HilingualTheme.colors.black,
-                modifier = Modifier.align(Alignment.CenterHorizontally)
-            )
-            Spacer(Modifier.height(24.dp))
-            Row(
-                horizontalArrangement = Arrangement.Start,
-                verticalAlignment = Alignment.CenterVertically,
-                modifier = Modifier
-                    .padding(12.dp)
-                    .fillMaxWidth()
-                    .noRippleClickable(
-                        onClick = onReportClick
-                    )
-            ) {
-                Icon(
-                    imageVector = ImageVector.vectorResource(DesignSystemR.drawable.ic_report_24),
-                    contentDescription = null,
-                    tint = Color.Unspecified,
-                    modifier = Modifier
-                        .size(24.dp)
-                )
-                Spacer(Modifier.width(8.dp))
-                Text(
-                    text = "신고하기",
-                    style = HilingualTheme.typography.bodySB14,
-                    color = HilingualTheme.colors.gray700
-                )
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Preview(showBackground = true)
-@Composable
-private fun FeedbackReportPreview() {
-    var isVisible by remember { mutableStateOf(true) }
-
-    HilingualTheme {
-        HilingualTheme {
-            FeedbackReportBottomSheet(
-                isVisible = isVisible,
-                onDismiss = { isVisible = false },
-                onReportClick = {}
-            )
-        }
-    }
-}
